boards: nrf: fix deprecated I2C properties
Commit 821c03a14a66942fb69c9cc056e2751bd0a8f258 ("i2c: nordic: switch
to phandle arrays for pinmux") deprecated some Nordic devicetree
properties.

When boards get merged with stale CI results (i.e. if CI results are
from a mainline commit earlier than 821c03a1), we will get deprecation
warnings, which twister treats as errors.

Play whack-a-mole with the ones that are in tree.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>