dts: st: f3: fix adc1 node
The adc1 of the stm32f373 uses an address space and a bus that are
different from the other microncontrollers of the f3 family. So, let's
remove the adc1 node from the stm32f3.dtsi and add it directly, with
respective correct values, in each of the SoC variant files.

Signed-off-by: Dario Binacchi <dariobin@libero.it>