boards: arm: mimxrt1020_evk: SD Card
Porting code from mimxrt1050_evk to this board to enable use of the SD
card.  Note that this board does not have the power pin connected to the
SD card and must use the no-1-8-v option for the USDHC driver as in
PR #34205 for the mm_swiftio.

Tested with samples/subsys/fs/fat_fs.

Signed-off-by: Jake Mercer <jake.mercer@civica.co.uk>