boards: arm: rcar h3ulcb: Add OpenOCD configuration
On R-Car the Cortex R7 is usually not the boot processor.
This configuration file make use of the Cortex A57 processor,
to initialize the Cortex R7.

It boils down to few steps:
 - Apply power to the Cortex R7
 - Set the boot address for the Cortex R7
 - loading a firmware
 - releasing the Cortex R7 reset

This configuration file also rely on A57 bootloaders,
to initialize device, memory, and security groups.

This file has been tested on openocd 0.10.0+dev-01508-gf79c90268-dirty,
shipped with zephyr sdk 0.12.4, and on openocd master
65c9653cc768f77a5e8cf2af73e0f40d614bdec2.

Thread awareness is possible thanks to this patch:
http://openocd.zylin.com/#/c/6369/

Signed-off-by: Julien Massot <julien.massot@iot.bzh>