net: ip: only define socklen_t if not already defined
There is a chance that a previous declaration exists of socklen_t
through the unistd.h header. Luckily that header defines a macro
when it defines the type. This commit therefore uses that define to
know if the type has already been declared or not, similar to what
unistd.h does.

Closes #57195

Signed-off-by: Hein Wessels <heinwessels93@gmail.com>